micro-services/signal-sheet-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
httpx==0.26.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.16.0
google-auth==2.27.0
numpy==1.26.4
Pillow==10.2.0
python-dateutil==2.8.2
